Can draw rect with fixed position, size and color
This commit fixes a bug due to `XuiRenderStatus` enum introduced in last
commit and adds code to render a rect2d with fixed position, size and
color.

The next step is to create a UI tree and pass that to shaders to be
rendered. Maybe we can try instanced render for this!