Updated Graphics Plugin README.md
I realized that nothing was wrong with the `draw_rect_2d` call, and it
was just my expectation that was wrong. On each new draw call the
framebuffer get's swapped and new content appears.

To fix this I either call the draw command multiple times so that all
framebuffers have same content or I draw complete UI in single draw
call.

From here on the effors would be to draw complete UI on single draw
call.