[WIP] Refactoring Start To Add Shapes.
This commit starts adding work for refactoring of "Vulkan Graphics
Plugin" to add support for rendering the whole UI in a single draw call.

We'll send shape data and vertex data before rendering and then
render everything, everywhere all at once!

The detailed explanation is given in the README.md of the plugin.